@NiklasNummelin 2023-11-20
Nice! I have a branch here also with some other optimizations I want to finish up ASAP https://github.com/EmbarkStudios/puffin/tree/perf-improvement-test