@TimonPost 2023-11-21
I'll have a look at removing unsafe access and maybe just going for the stream-info-once solution as proposed in #167